Link detach crash and memory leak fixes
Fixes the crash when calling link_destroy after message_sender_destroy thereby callback not being cleared, and calling into deallocated memory.  Also fixes an issue whereby messages that are not being send due to busy are not removed again from the link, which causes memory leak and worse, crashes in case the send was already notified as error or timeout but still remains on the link and then completes later.